Add Mercury Logo to Sponsors
Add the mercury logo to the sponsors section. This addition caused the
contents of the footer to overflow onto a second line. The way that we
were setting the width of the component seemed to cause the contents
of the footer on the second row to disappear at desktop resolutions,
so we now set min and max width, which allows the contents to span
multiple lines gracefully.